fix(ConfirmProposedTx): pass necessary props to ConfirmTxDetails and further down to SignOrExecuteFormV2
